disasm: avoid gcc warning

+++ disasm.c
@@ -107,7 +107,7 @@ const char *decode_operand(uint64_t bits)
 			    decode_fp20(embedded_consts >> offset));
 			embedded_consts_used = 1;
 		} else
-			dst += sprintf(dst, "#%d", (bits >> 5) & 1);
+			dst += sprintf(dst, "#%d", (int)(bits >> 5) & 1);
 	} else {
 		/* normal registers */
 		dst += sprintf(dst, "%c%c%d",
